projects/polymarket-bot/requirements.txt
py-clob-client>=0.5.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23.0
